requirements.txt
# AI和机器学习
transformers>=4.30.0
diffusers>=0.20.0
torch>=2.0.0
accelerate>=0.20.0
xformers>=0.0.20

# 图像处理
Pillow>=9.5.0
opencv-python>=4.8.0
imageio>=2.31.0
numpy>=1.24.0

# 音频处理
pydub>=0.25.1
librosa>=0.10.0
gTTS>=2.3.0
pygame>=2.5.0

# 视频处理
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Web和API
requests>=2.31.0
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 工具库
tqdm>=4.65.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 系统工具
psutil>=5.9.0
GPUtil>=1.4.0
